Huge commit. Brought codebase up-to-speed under Cocoa and removed the Wizard-based materialize.

+++ cocoa/AddLocationWindow.m
@@ -109,12 +109,12 @@ http://www.hardcoded.net/licenses/hs_license
 }
 
 //Data Source
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     return [_mediaList count];
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
     return [self locationNameFromPath:[_mediaList objectAtIndex:rowIndex]];
 }
